Agentic_Rag_Application/pyproject.toml
[tool.poetry]
name = "agentic-rag-application"
version = "0.1.0"
description = ""
authors = ["Jawad Ahmed <jawad.kohat2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
llama-index = "0.10.23"
llama-index-core = "0.10.68"
pandas = "2.2.2"
scikit-learn = "1.4.2"
tabula-py = "2.9.3"
chromadb = "0.4.24"
llama-index-embeddings-huggingface = "0.1.4"
camelot-py = "0.11.0"
pypdf2 = "^3.0.1"
llama-index-vector-stores-chroma = "0.1.9"
pymupdf = "1.24.2"
jpype1 = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
